made _assert_spatialdata_objects_seem_identical() into a util
