+++ src/spatialdata/_utils.py
@@ -5,12 +5,15 @@ import re
 import warnings
 from collections.abc import Generator
 from copy import deepcopy
-from typing import Any, Callable, TypeVar, Union
+from typing import TYPE_CHECKING, Any, Callable, TypeVar, Union
 
 import numpy as np
 import pandas as pd
 from anndata import AnnData
+from anndata.tests.helpers import assert_equal
 from dask import array as da
+from dask.dataframe import DataFrame as DaskDataFrame
+from dask.delayed import Delayed
 from datatree import DataTree
 from geopandas import GeoDataFrame
 from multiscale_spatial_image import MultiscaleSpatialImage
@@ -24,6 +27,9 @@ from spatialdata.transformations import (
     get_transformation,
     set_transformation,
 )
+
+if TYPE_CHECKING:
+    from spatialdata import SpatialData
 
 # I was using "from numbers import Number" but this led to mypy errors, so I switched to the following:
 Number = Union[int, float]
@@ -311,3 +317,42 @@ def _error_message_add_element() -> None:
         "write_labels(), write_points(), write_shapes() and write_table(). We are going to make these calls more "
         "ergonomic in a follow up PR."
     )
+
+
+def _assert_elements_left_to_right_seem_identical(sdata0: SpatialData, sdata1: SpatialData) -> None:
+    for element_type, element_name, element in sdata0._gen_elements():
+        elements = sdata1.__getattribute__(element_type)
+        assert element_name in elements
+        element1 = elements[element_name]
+        if isinstance(element, (AnnData, SpatialImage, GeoDataFrame)):
+            assert element.shape == element1.shape
+        elif isinstance(element, DaskDataFrame):
+            for s0, s1 in zip(element.shape, element1.shape):
+                if isinstance(s0, Delayed):
+                    s0 = s0.compute()
+                if isinstance(s1, Delayed):
+                    s1 = s1.compute()
+                assert s0 == s1
+        elif isinstance(element, MultiscaleSpatialImage):
+            assert len(element) == len(element1)
+        else:
+            raise TypeError(f"Unsupported type {type(element)}")
+
+
+def _assert_tables_seem_identical(sdata0: SpatialData, sdata1: SpatialData) -> None:
+    tables0 = sdata0.tables
+    tables1 = sdata1.tables
+    assert set(tables0.keys()) == set(tables1.keys())
+    for k in tables0:
+        t0 = tables0[k]
+        t1 = tables1[k]
+        assert_equal(t0, t1)
+
+
+def _assert_spatialdata_objects_seem_identical(sdata0: SpatialData, sdata1: SpatialData) -> None:
+    # this is not a full comparison, but it's fine anyway
+    assert len(list(sdata0._gen_elements())) == len(list(sdata1._gen_elements()))
+    assert set(sdata0.coordinate_systems) == set(sdata1.coordinate_systems)
+    _assert_elements_left_to_right_seem_identical(sdata0, sdata1)
+    _assert_elements_left_to_right_seem_identical(sdata1, sdata0)
+    _assert_tables_seem_identical(sdata0, sdata1)
